google-search-agent/requirements.txt
google-adk
google-genai>=1.56.0
google-cloud-aiplatform[reasoningengine]
google-auth
